refactor: reduce number of actions and avoid processing in reducers (#401)
# Issue Number: #379 

# Body

This refactors the actions so that there are fewer actions and the
reducers only update the store. No modifying the data on the way in. The
data modification is now done with a class that gets call, processes the
data, and then dispatches and action to update the store.

In the future, more of the code will follow this pattern which will lend
itself well to using Signals instead of the Redux store.