refactor(smart-ngrx): remove the distinction between virtual and eager arrays (#572)
# Issue Number: #571

# Body

The code now checks to see if the array is an array. If it is, it uses
the old way of dealing with child arrays. If not, it assumes it is a
virtual array.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **New Features**
- Simplified department structure by removing nested `children`
properties from various department definitions, facilitating a flatter
data model.
- Enhanced clarity and performance in the handling of parent-child
relationships in the application.

- **Bug Fixes**
- Improved robustness of functions related to entity selection by
modifying parameter inputs and removing unnecessary variables.

- **Tests**
- Restructured the test cases for better maintainability following
updates to the function parameters and logic.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->